Convert indices to size_t in decomposition

+++ src/MatrixDecompositions.cpp
@@ -3,7 +3,7 @@
 
 void MatrixDecompositions::LU(const Matrix& A, Matrix& L, Matrix& U)
 {
-    unsigned int n = A.Rows();
+    size_t n = A.Rows();
 	if (n != A.Columns()) throw std::invalid_argument("LU Decomposition requires square input");
 
 	// ALGORITHM 3.2.1 FROM MATRIX COMPUTATIONS, 4TH ED
@@ -11,13 +11,13 @@ void MatrixDecompositions::LU(const Matrix& A, Matrix& L, Matrix& U)
     L = Matrix::Identity(n);
 	U = Matrix(A);
 
-    for (unsigned int k = 0; k < n - 1; k++)
+    for (size_t k = 0; k < n - 1; k++)
     {
-        for (unsigned int i = k + 1; i < n; i++) U(i, k) /= U(k, k);
+        for (size_t i = k + 1; i < n; i++) U(i, k) /= U(k, k);
 
-        for (unsigned int i = k + 1; i < n; i++)
+        for (size_t i = k + 1; i < n; i++)
         {
-            for (unsigned int j = k + 1; j < n; j++)
+            for (size_t j = k + 1; j < n; j++)
             {
                 U(i, j) -= U(i, k) * U(k, j);
             }
